requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
Django==4.2.2
django-extensions==3.2.3
django-webpush==0.3.5
gunicorn==20.1.0
http-ece==1.1.0
idna==3.4
kombu==5.3.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
ngrok==0.8.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
prompt-toolkit==3.0.38
py-vapid==1.9.0
pycparser==2.21
python-dateutil==2.8.2
pywebpush==1.9.4
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
